+++ plugin-dev/Source/Sentry/Private/HAL/PlatformSentrySamplingContext.h
@@ -2,11 +2,10 @@
 
 #pragma once
 
+#if PLATFORM_ANDROID
+#include "Android/AndroidSentrySamplingContext.h"
+#elif PLATFORM_APPLE
+#include "Apple/AppleSentrySamplingContext.h"
+#else
 #include "Null/NullSentrySamplingContext.h"
-
-static TSharedPtr<ISentrySamplingContext> CreateSharedSentrySamplingContext()
-{
-	// Sampling context is supposed to be created internally by the SDK using the platform-specific implementations.
-	// Currently, it doesn't provide default constructor for Apple/Android thus we can only return Null-version here.
-	return MakeShareable(new FNullSentrySamplingContext);
-}
+#endif